+++ pkg/tfutil/reshape.go
@@ -7,15 +7,6 @@ import (
 	"github.com/tensorflow/tensorflow/tensorflow/go/op"
 )
 
-func shapeToint64Shape(shape []int) []int64 {
-	newShape := make([]int64, len(shape))
-	for i := range shape {
-		newShape[i] = int64(shape[i])
-	}
-
-	return newShape
-}
-
 // Reshape reshapes to new shape
 func (g *Tensor[T]) Reshape(shape ...int) error {
 	x, err := g.Marshal()
@@ -44,29 +35,23 @@ func (g *Tensor[T]) Reshape(shape ...int) error {
 		),
 	)
 
-	// Define the operation node that accepts A & x as inputs
-	product := op.Reshape(root, X, Y)
+	// define reshape operation
+	O := op.Reshape(root, X, Y)
 
 	graph, err := root.Finalize()
 	if err != nil {
 		return fmt.Errorf("failed to import graph: %w", err)
 	}
 
-	// prepare data feed specifying names of the operation.
-	// names x and dim come from python code, see def of reshape
-	// function taking inputs x and dim
 	feeds := map[tf.Output]*tf.Tensor{
 		X: x,
 		Y: y,
 	}
 
-	// prepare data outputs from tensorflow run.
-	// Identity is the final output point of the graph.
 	fetches := []tf.Output{
-		product,
+		O,
 	}
 
-	// start new session
 	sess, err := tf.NewSession(
 		graph,
 		&tf.SessionOptions{},
@@ -76,7 +61,6 @@ func (g *Tensor[T]) Reshape(shape ...int) error {
 	}
 	defer sess.Close()
 
-	// run session feeding feeds and fetching fetches
 	out, err := sess.Run(feeds, fetches, nil)
 	if err != nil {
 		return fmt.Errorf("failed to run tf session: %w", err)
