+++ pkg/tfutil/scale.go
@@ -45,29 +45,23 @@ func (g *Tensor[T]) Scale(value T) error {
 		),
 	)
 
-	// Define the operation node that accepts A & x as inputs
-	product := op.Mul(root, X, Y)
+	// define multiplication operator
+	O := op.Mul(root, X, Y)
 
 	graph, err := root.Finalize()
 	if err != nil {
 		return fmt.Errorf("failed to import graph: %w", err)
 	}
 
-	// prepare data feed specifying names of the operation.
-	// names x and dim come from python code, see def of reshape
-	// function taking inputs x and dim
 	feeds := map[tf.Output]*tf.Tensor{
 		X: x,
 		Y: y,
 	}
 
-	// prepare data outputs from tensorflow run.
-	// Identity is the final output point of the graph.
 	fetches := []tf.Output{
-		product,
+		O,
 	}
 
-	// start new session
 	sess, err := tf.NewSession(
 		graph,
 		&tf.SessionOptions{},
@@ -77,7 +71,6 @@ func (g *Tensor[T]) Scale(value T) error {
 	}
 	defer sess.Close()
 
-	// run session feeding feeds and fetching fetches
 	out, err := sess.Run(feeds, fetches, nil)
 	if err != nil {
 		return fmt.Errorf("failed to run tf session: %w", err)
